security test for release-6.5

+++ tests/integration_tests/bank/run.sh
@@ -9,6 +9,10 @@ CDC_BINARY=cdc.test
 SINK_TYPE=$1
 
 function prepare() {
+	if [ "$SINK_TYPE" == "kafka" ]; then
+		return
+	fi
+
 	rm -rf $WORK_DIR && mkdir -p $WORK_DIR
 
 	start_tidb_cluster --workdir $WORK_DIR
@@ -25,16 +29,15 @@ function prepare() {
 }
 
 trap stop_tidb_cluster EXIT
-# No need to support kafka and storage sink.
-if [ "$SINK_TYPE" == "mysql" ]; then
+# kafka is not supported yet.
+if [ "$SINK_TYPE" != "kafka" ]; then
 	prepare $*
 
 	cd "$(dirname "$0")"
 	set -euxo pipefail
 
 	GO111MODULE=on go run bank.go case.go -u "root@tcp(${UP_TIDB_HOST}:${UP_TIDB_PORT})/bank" \
-		-d "root@tcp(${DOWN_TIDB_HOST}:${DOWN_TIDB_PORT})/bank" --test-round=20000 \
-		-a "${DOWN_TIDB_HOST}:${DOWN_TIDB_STATUS}"
+		-d "root@tcp(${DOWN_TIDB_HOST}:${DOWN_TIDB_PORT})/bank" --test-round=20000
 
 	cleanup_process $CDC_BINARY
 	echo "[$(date)] <<<<<< run test case $TEST_NAME success! >>>>>>"
